requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
pytesseract==0.3.13
pillow==11.2.1
python-dotenv==1.1.0
requests==2.31.0
